scap: Stop cloning over /p/
This url is broken from gerrit 2.16 (and also completely optional)

Take in account an existing repository might still have /p/ so we can
still extract the repository basename. A better fix would be to have the
provider to update the local copy to use the canonical URL.

Add some unit tests written by Antoine to cover the scap_source
provider.

Bug: T218844
Signed-off-by: Antoine Musso <hashar@free.fr>
Change-Id: Iad6fe0d1d7a6e3917e23cb49addbdd609275fa98