Revert "sudo: decouple sudo from sudo-ldap"
It turns out that for some reason we are including raid stuff on all CloudVPS
VMs. The raid stuff creates a sudo user, with the now-default 'sudo' flavour.

We would need to either override that sudo selection by means of profiles or
totally reconsider my approach. Anyway this doesn't work, so reverting.

This reverts commit 5f0c714683db6d9ae38e799e9482c911bc61c815.

Bug: T221225
Change-Id: Ida04144cd704bf21af69b8fd293eeacb4e7044b5
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>