base: make RAID optional
This patch makes the RAID check optional, by means of a new class parameter.
At the profile level, the default is to include it (behaviour unchanged) but
you have a new hiera key 'profile::base::check_raid' that can override the
behavior.

The base module is included in every node, but the RAID config does not make
sense in VMs, specially in CloudVPS (AKA labs). The hiera config introduced
by this patch disables RAID checks for VMs.

Change-Id: I530488f090b06d526fcc078bd09747c9fa9b4f9a
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>