smart-data-dump: add '-l error' to facter command to suppress warnings
This patch updates the smart-data-dump script to suppress warning
messages if the major version facter is 3.  Further investigation will
be conducted on the underlining warning

Currently we are getting a lot of cron spam from cp* machines
because of a bug in facter.

It spits out warnigns if it fails to parse the output of
'/sbin/ip -6 route show'.

Very similar but not identical to

https://tickets.puppetlabs.com/browse/FACT-1394

where it fails to parse if "linkdown" appears in the output.

That is not the case for us but:

[cp3041:~] $ sudo facter --puppet --json raid | grep WARN

WARN  puppetlabs.facter - Could not process routing table entry:
Expected a destination followed by key/value pairs,
got '2620:0:861:107:10:64:48:101 via fe80::1 dev eno1 metric 1024  mtu lock 1450 pref medium'

Running facter with -d shows debug output to be:

DEBUG leatherman.execution:93 - executing command: /sbin/ip route show

Changing the log level to keep errors but ignore warnings.

Bug: T86552
Change-Id: Icfb4d97a0b81c5396f3d3e31a2763e82a24fe842