sudo: decouple sudo from sudo-ldap
This is the 3º patch attempt. First 2 were reverted because:
* attempt #1 produced some broken CloudVPS VMs
* attempt #2 was merged with not enough review

Callers of the sudo module can choose which sudo flavor they want:

* sudo (normal sudo, the default)
* sudo-ldap (a LDAP/nscd/nslcd based sudo)

The sudo-ldap flavor is usually used in CloudVPS, but if using sssd rather
than nscd/nslcd we need sudo instead of sudo-ldap.

This patch involves decoupling the sudo class into another subclass called
sudo::sudoldap, which contains the specific bits related to that version of
sudo.

The desired outcome is that a CloudVPS VM can choose which LDAP client stack to
use between 2:
* classic, using nscd/nslcd, and sudo-ldap
* sssd, using the sssd daemon and sudo (no sudo-ldap)

Both flavors are incompatible.

A new global hiera key is introduced: 'sudo_flavor', which defaults to
'sudoldap' in CloudVPS and to 'sudo' in prod. Eventually, this hiera key will
be dropped, since we aim to a near future in which we run sssd in CloudVPS and
we can just use 'sudo' everywhere.

Bug: T221225
Change-Id: I5481218bd044fa6c46d8a66cd60e9d58b339e200
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>