mariadb: Chmod /srv/backups/dumps o+x so the disk space check works
du requires +x permissions on the parent directories for it to work
(it is used by the disk space check). Because we currently mount
the ssd disks on /srv/backups/dump/ongoing, add +x privileges (not
read or write, so it will have no effect on data access) to other
and group. We must do it for all its parents recursively, so
/srv/backups and /srv/backups/dumps .

This will solve the CRIT for Disk space alert "DISK CRITICAL -
/srv/backups/dumps/ongoing is not accessible: Permission denied"

Bug: T219399
Bug: T218336
Change-Id: Id12dfd4b10d6afa22e4c38b868ac6188d53c681f