mariadb-backups: Force snapshots in certain order
With the current setup, backups happen in random order.
This has some advantages, like not doing things always
in the same way, but it is higly inefficient (hosts are
idle while others are too busy).

With the intention of optimizing resources, so not 2 backups
happen frequently on the same source instance or same target host,
specify an intended order on the sections to be backed up.

Because YAML and python dictionaries are not ordered in nature,
use a separeate key/option 'order' and start those in that
particular order. If an order is not specified, do them after
the ordered ones, in random order.

Bug: T206203
Change-Id: Ibd43a5e783c7f406de15b90eeb10f560a0da5eef