upload_ats: include ats-be in backend_services
confd clients fails if there is no data under a given etcd key. For
example, if no upload host has been converted to ATS in codfw yet,
reading the following key would fail:
/conftool/v1/pools/codfw/cache_upload/ats-be

For this reason we are setting the hiera value used to decide which keys
to watch, profile::cache::varnish::frontend::backend_services, on a
per-role/per-DC basis when at least one node in that role/DC has been
converted (see commit 6984168). With that being said, if
role::cache::upload_ats is applied on a given node, that means that
there is at least one ATS node in the DC/role of the node. Hence,
include 'ats-be' regardless of the DC for role upload_ats.

Bug: T222937
Change-Id: I8976e8faa6af43c22b1954eaa0c3e3422590c47b