toolforge: remove memory overcommit limitations on cron host
The cron host for toolforge dubiously requires the exec_environ setup.
Unfortunately, that limits its overcommit memory by a bit.  Since this
*should* just be launching jobs, overcommits should be transient and
not actually disruptive on this host.  Therefore, I'm refactoring it out
and including it in the roles of the exec hosts and the bastions (where
the setting is much more likely to be required).

Bug: T222255
Change-Id: Ic4e3238b68196080b578c31696e9e7881974e23c