ATS: require explicit Cache-Control/Expires
With the current caching settings it is hard to figure out whether a
certain response will be cached, and impossible to say exactly for how
long.

Instead of relying on heuristics and Negative Response Caching when
Cache-Control is missing, make sure that only responses with explicit
values set for Cache-Control/Expires are cached. Set Cache-Control in
Lua for responses that need to be cached, and unset it for responses
that must not.

Bug: T222937
Change-Id: I027627b32757d997675ea9a92065b50d7eb057ae