openstack: in stretch don't use libjs-jquery from openstack-mitaka-jessie
The version of libjs-jquery provided in stretch is good enough for us to don't
need the one in openstack-mitaka-jessie.
Moreover, this avoids the downgrade situation in stretch systems.

This package is required in basically all servers, because nova-common:

=== 8< ===
aborrero@cloudvirt1024:~$ aptitude why libjs-jquery
i   nova-common     Depends python-nova (= 2:13.1.0-2~bpo8+1)
i A python-nova     Depends python-routes
i A python-routes   Depends libjs-sphinxdoc (>= 1.0)
i A libjs-sphinxdoc Depends libjs-jquery (>= 1.11.1)
=== 8< ===

Previous to this patch, on a jessie system:

=== 8< ===
libjs-jquery:
  Installed: 1.11.3+dfsg-4~bpo8+1
  Candidate: 1.11.3+dfsg-4~bpo8+1
  Version table:
 *** 1.11.3+dfsg-4~bpo8+1 0
       1001 http://apt.wikimedia.org/wikimedia/ jessie-wikimedia/openstack-mitaka-jessie amd64 [..]
        100 /var/lib/dpkg/status
     1.7.2+dfsg-3.2+deb8u6 0
        500 http://security.debian.org/ jessie/updates/main amd64 Packages
     1.7.2+dfsg-3.2 0
        500 http://mirrors.wikimedia.org/debian/ jessie/main amd64 Packages
=== 8< ===

Previous to this patch, on a stretch system:

=== 8< ===
libjs-jquery:
  Installed: 3.1.1-2+deb9u1
  Candidate: 1.11.3+dfsg-4~bpo8+1
  Version table:
 *** 3.1.1-2+deb9u1 500
        500 http://mirrors.wikimedia.org/debian stretch/main amd64 Packages
        100 /var/lib/dpkg/status
     1.11.3+dfsg-4~bpo8+1 1001
       1001 http://apt.wikimedia.org/wikimedia jessie-wikimedia/openstack-mitaka-jessie amd64 [..]
=== 8< ===

After this patch, on a stretch system, the libjs-jquery package from
openstack-mitaka-jessie will have a pinning priority of -1, thus won't be
installed and the native stretch version will be the candidate.

Bug: T222862
Change-Id: Ie1e25cddc70078262d119c54036a4891d3d6196d
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>