Switch deployment-prep to facter 3 / puppet 5
Currently labs at large is still at facter 2 / puppet 4 (while production
is almost done migrating to facter 3 / puppet 5). Given that beta is
supposed to mimic production it makes sense to also use it here. It
will also allow us to test the updated packages for puppet-master,
puppet-master-passenger and puppetdb which are currently held back in
production. And it will be a useful test case to also migrate the rest
of labs at a later point.

Bug: T219803
Change-Id: I15a716353de92355ec1a8a72e5a5c02d77b0b0a7