icinga: check_eth: fix bridge and tap interface regexp
Previous to this patch, the regexp won't match interfaces with names like this:
 * brq7425e328-56
 * tapbcf16ed2-63

So, switch to a even more general regexp in which we match anything that starts
with "tap" or "br".

This is fixing a recent change in puppet (related to the facter version
perhaps) in which we saw this change:

[..]/content) -for INTERFACE in enp4s0f0 enp4s0f1 lo ; do
[..]/content) +for INTERFACE in brq7425e328-56 enp4s0f0 enp4s0f1 lo tapbcf16ed2-63 ; do

So this patch should just revert that previous change.

Bug: T223107
Change-Id: I062a113356358ed12face26a2f6ff6eb1f2495ec
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>