openstack: fix weird systemd state in standby controller node
The standby controller node starts the openstack services by default due to
debian package installation.
But we don't want to run them in the standby node, so we must ensure they are
disabled via puppet.

There is a race between service startup and puppet ensuring they are stopped:
the services may be already started before puppet agent is run. The problem
is that firewal doesn't allow rabbitmq connections from the standby node to the
active one, which causes the services to fail startup. This will produce an
icinga warning that puppet can't solve, because the service is stoppped (only
in fail state).

To address this, my approach is to:
 * let's enable rabbitmq connections from the standby to the active node
 * so puppet agent can gracefully shutdown the openstack services
 * ensure neutron-server is stopped in the standby node (missing)

Bug: T215544
Change-Id: Ia7021148cccfea4f95d7839b5e7885c32a5dc738
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>