Empty mediawiki_memcached_servers for 3 mw hosts
This is a test to see if nutcracker is still used for
memcached. The configuration is removed for an appserver,
a api appserver and a jobrunner. The regex.yaml solution
seems a quick and easy solution to provide a testing environment,
since the value is added to eqiad|codfw.yaml and cannot be easily
overridden via the hiera's canary role.

Bug: T214275
Change-Id: I253b3b29450a827a1fd43ca50aea35fe4313a676