debug_proxy: force http/1.1 when proxying
By default nginx uses http/1.0 when proxying, and when
we are using X-Wikimedia-Debug, our requests go to mwdebug*
via an nginx.

In this case, when we deliberately sent a request that
would return 500 to mwdebug*, HHVM on mwdebug would
return the 500 error but with 'Content-length: 0'. In turn,
Varnish would server the user a Varnish error and not
the PHP error.

PHP7 returns the correct content length when using http/1.0.
(replaces Ibe68d0b6ddc26e295212e7d6ca3c0287a2669826)

Bug: T217846
Change-Id: I8687896a0b36619eac1c608f01dab75eda9e3fe5