wiki replicas: Remove reference to old user fields
This makes the following user-visible changes, which should be
announced:

* archive: Remove ar_user and ar_user_text
* archive_userindex: Remove ar_user and ar_user_text
* filearchive: Remove fa_user and fa_user_text
* filearchive_userindex: Remove fa_user and fa_user_text
* image: Remove img_user and img_user_text
* ipblocks: Remove ipb_by and ipb_by_text
* ipblocks_ipindex: Remove ipb_by and ipb_by_text
* logging: Remove log_user and log_user_text
* logging_logindex: Remove log_user and log_user_text
* logging_userindex: Remove log_user and log_user_text
* oldimage: Remove oi_user and oi_user_text
* oldimage_userindex: Remove oi_user and oi_user_text
* recentchanges: Remove rc_user and rc_user_text
* recentchanges_userindex: Remove rc_user and rc_user_text
* revision: Remove rev_user and rev_user_text
* revision_userindex: Remove rev_user and rev_user_text

To attempt to help with T221339, this also removes a few COALESCE()
calls. This could in theory result in some xx_actor fields being
returned as NULL rather than 0, but in practice that should never
happen.

It also adds extra join conditions for revision + revision_actor_temp to
attempt to allow the planner to better know when it can propagate
conditions (even though it's likely to be ineffective for
rev_timestamp/revactor_timestamp, see Iad6c0c2f).

This also updates the compat views for the above tables to stop
referring to the named fields in the underlying tables.

This does not attempt to convert LEFT JOINs into INNER JOINs, to avoid
any potential repeat of T212972#5171372. I'll leave that for another
patch.

Bug: T223406
Bug: T221339
Change-Id: Id6dd77305fe195a3eec2e9ccae8c3f1bc6f5c08d