phabricator: add vcs listen addresses for phab1003
Add the missing extra VCS listen addresses for phab1003.

As previously used on phab1002 and then renamed in DNS.

Note the unmapped IPv6 address is removed, double check it's not needed.

Also the commented IP is the git-ssh service name that needs to be
switched on the day of moving phab1001 to phab1003.

Bug: T190568
Change-Id: I6f572e450df02ca5ff204dee147780783a2d2e8f