Expose cloudelastic to wmfcloud
Adds a new tlsproxy port to cloudelastic machines that only allows
read requests. This port is then exposed via ferm to labs networks,
allowing read-only access to cloudelastic from wmfcloud.

For the initial implementation this is limited to GET requests to
ensure only reads are accepted. Possibly this will expand to other
options, such as POST to _msearch, but generalizing that into the
tlsproxy nginx template will happen after more thought is given to an
appropriate implementation.

Bug: T223519
Change-Id: I12e0118674ab76edc872e47cf4d60419a804eea1