wiki replicas: remove the coalesce guard around revactor_actor
This is only for the revision_userindex table since it's the table most
likely to benefit from an index that isn't working with that in the way.
It is possible this will allow NULL to be returned for an integer field,
but the backfilling should have taken care of that at this point.

Bug: T221339
Change-Id: Idf27d050457e8259e8d95b205e5162122271d05a