tlsproxy::localssl: Don't include tlsproxy::ocsp if we're not doing OCSP
With tlsproxy::ocsp, sslcert::ocsp::hook will be used, which will require
sslcert::ocsp::init, which will set up update-ocsp-all, but that won't work if
acme_chief::cert has not been told to use OCSP, because the
/etc/update-ocsp.d/${title}-${version}-${type}.conf file will not have been
created, and update-ocsp-all cannot handle that.

I'm guessing that this is the source of T223734

PS2: This is equivalent to wrapping acme-chief's tlsproxy::ocsp in an if check
for $do_ocsp.

Bug: T223734
Change-Id: Iebf31987836ad041b7172af7ece770b96c4ba72d