ldap::client::utils: don't require python-ldap if already declared
If the package resource is already declared elsewhere, don't require it.
Then, I introduce a package resource ensuring is absent in the openstack
code tree.

This is required in stretch, where mitaka keystone requires python-pyldap
rather than python-ldap, which is a compatible fork by the time of stretch.

I know this is not a very elegant solution, but can serve as workaround until
we move past openstack mitaka.

You can read further details in this commit:
 9a9e83c6241d65ccca54b9c49dc36468bdd17fa9

Bug: T223832
Change-Id: I4e85d0a85fe378832a11d1c0b25cf47184d4cb91
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>