gerrit: allow customizing LDAP config in Hiera
A different approach to achieving what was attempted
in I99c15ca596de8e8e2b.

Allow customizing the LDAP config in Hiera.

But instead of merging 2 different config sources keep it simpler
by allowing to put the entire $ldapconfig hash in Hiera.

This should then be set in project Hiera.

If it's not found in Hiera, do everything exactly as before and
include the global labs LDAP config.

Change-Id: I80b0b1e404208d2f0c90ea24b1a6f29dcdd26708