mediawiki::php: bump the APC shm size on php7
We've had a relatively small APCu shm size up to now, as during the
pilot phase of the migration we didn't notice any APC exhaustion
ongoing.

We're now in a very different situation, also it's possible frequent
expiries in APC would cause some issues we've been seeing lately with
excessive requests of data to the backends.

Bug: T223310
Change-Id: I8b87065b0912e80c23f6e7910bc48b95b03356ed