Add gnupg to d-i base installer for buster
Preseeding a local repository via apt-setup/localX/repository is broken
in Buster, as apt-key depends on gnupg, but it gets no longer installed
by default unlike in previous releases.

I've submitted a merge request on Salsa to fix apt-setup, but it's stalled
for a month. To workaround this bug, let's pull in gnupg explicitly in
d-i.

Bug: T213527
Change-Id: I2053748f1c4d5f267c5bf5c822655722d686ffc5