transparency report: remove git cloning of private site
The git cloning of the private site recently stopped working due
to changes on the Gerrit side. A regular Gerrit user was used to
clone the repo.

But we also don't seem to need this site anymore so removing it
rather than dealing with how to reactivate it properly with a bot user.

Change-Id: Iec6599d29559e2dab505760aa037b6762f78980f