pybal: remove redundant hiera config
The config in theses files does not apply as intended due to the fact
that the common tree is first in the hierarchy and has a value.

As this is a no-op i would like to remove it however im hopping to
change the hiera priority so that this config would be valid.  At which
point i propose to re-add this config


Hosts: pybal-test2001.codfw.wmnet
Related: I8beafbbb40aad76e7a7cdee4d336cc3944eb1ef8

Change-Id: Id46e95736001582f2fdad42b83980ca3d12a1bd9