wiki replicas: Improve index usage for queries against revision_userindex
For the time being, the data for the `rev_actor` field in MediaWiki is
actually stored in a separate table (`revision_actor_temp`) because
altering `revision` was expected to be far too expensive an operation in
Wikimedia production. The introduction of the `actor` and `comment`
tables is in fact a significant part of trying to make that *not* be the
case.

To allow MediaWiki's own queries to still be able to use sane indexes,
the `revision_actor_temp` table includes denormalized copies of
`rev_timestamp` and `rev_page` to support the intended future indexes
combining `rev_actor` and those fields.

The WMCS replica views are using joins to expose `rev_actor` as if it
were part of the `revision` table, to avoid forcing tool authors to
change their code to use `revision_actor_temp` only to soon afterwards
change again when that table is removed. But convincing MariaDB that
queries via the view using `rev_actor` and `rev_timestamp`/`rev_page`
can use the denormalized fields turns out to be a bit tricky, since
Wikimedia wikis created before 2012 (i.e. most of them) have mismatched
types for `rev_timestamp` and `revactor_timestamp`.

But we probably don't have to actually worry about that. WMCS already
has a `revision_userindex` view that's intended for use when you're
making a query that wants to depend on the user (i.e. `rev_actor`). We
can probably get away with changing these views to use
`revision_actor_temp`'s denormalized fields instead of `revision`'s
versions, and then queries should naturally use the indexes on
`revision_actor_temp`! The down side is that this will prevent existing
queries using `rev_user` or `rev_user_text` in combination with
`rev_timestamp`/`rev_page` from using the best indexes, making them slow
instead, but since those fields are going away in a few weeks anyway
perhaps we can get away with not caring.

Note that giving the same treatment to `revision_compat` probably won't
work, since speeding up the `rev_actor` queries in this way would also
slow down queries against `rev_timestamp`/`rev_page` that don't use any
of the user/actor fields. It's probably best to just leave that view
as-is, particularly since the _compat views aren't really intended for
performance in the first place.

Bug: T221339
Change-Id: Iba2db780fb4ef6fbc1d5812d4d3dd5ed44675311