profile::mediawiki::php: make apc size configurable, bump for appservers
As clarified on the task, we still dedicate way less RAM to apc for php
than what we're doing for HHVM. Start progressively increasing it to
find the optimum balance of size and cache-hit ratio. Also, make it
separately configurable as on jobrunners and / or in beta we don't need
such a large memory space.

Bug: T223180
Change-Id: I5f276cb4dd6cbc8308527999193667024b8d35a9