phabricator: stop paging SRE for process checks, keep for https
Stop paging the entire SRE team for checks like "phd" process
not running. This happened during our most recent maintenance
but it was determined that "phd not running' is not even that
critical to justify that.

Remove the 'sms' contact group (despite the name this does not
really say anything about notification methods configured with
individual contacts but it means 'all of SRE') from most checks.

Only keep it for the https check in case phabricator is not reachable
at all that should probably still page everyone.

Bug: T224205
Change-Id: I2b960a730b24c3c43bad6e74fc893dd3d79ca492