toolforge: migrate k8s code from toollabs
This patch introduces all the puppet code for kubernetes, migrated from the
old toollabs tree. This new code in the toolforge tree will be used as a
new and improved kubernetes cluster for the Toolforge service.

The new role is meant to be used in Debian >= stretch.

Some bits can't be copied as they are, for example those bits related
to custom admission controllers, see T215678 for reference.
That will be reworked in a future iteration.

Bug: T215531
Change-Id: Ic7cec0696a01a948c61a0fcb8d9cf07b4fad82b0
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>