openstack: designate: mitaka: stretch: use pdns 3.x
Openstack mitaka doesn't support pdns 4.x, so we are forced to use 3.x,
which is present only in Debian Jessie. Introduce apt repo/pinning
config to force installation of the jessie package in stretch, using
archives.debian.org as repo, since jessie is already archived.

I'm explictly avoiding to introduce this stuff next to our other setup
related to jessie-wikimedia/openstack-mitaka, because that one is present
in every server, and this hack is only required briefly in cloudservices
nodes.

Not soon enough, we will be moving past Mitaka, and this won't be neccesary.

Bug: T224354
Change-Id: I3219228a1a47b0acaf5ad0323e3c7469397cbc56
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>