cassandra: Support client IPs in ferm
Up to now cassandra clusters made the assumption that no non-localhost
clients would connect to them. This is changing with the sesssionstore
service. Add support to the profile for allowing access to selected IPs.
Then just for the dev_cluster and the sessionstore cluster add the
kubernetes pod IPs. The kubernetes Pods are also going to be having
outgoing firewall rules allowing connection only for selected pods,
namely the sessionstore service

Bug: T220401
Change-Id: Iedef777a46acb06ea0ce974496974d9686334576