sssd: include the /etc/ldap.yaml file
This file is used by several scripts to contact the LDAP server.
One of the most important scripts in the one used to lookup for user
sshkeys, which is configured in the sshd_config.

Nowadays, sssd provides a native mechanism for that lookup, but it worth
keeping this file around just so we don't break more stuff and we can
eventually to a graceful transition to more sssd native mechanisms.

Bug: T224558
Change-Id: I88148ec2cd815c221deb882d6a827b2b0acdfbe7
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>