varnish: cache_upload miss/pass rate limit
Introduce global rate limiting for miss/pass traffic on cache_upload.
Begin with 1000/10s (100/s long term, with 1000 burst), which is what we
use on cache_text as well for the MW API, RestBase, and such and seems
to be working fine.  We could revise such values in the future.

Rate limit requests violating the UA policy more aggressively.

Bug: T224884
Bug: T224891
Change-Id: I5332e323643488864424c1f9881f5efdb47f42d6