wiki replicas: Add specialized views of the "comment" table
The comments view is not aware of the type of entity that a given query
is interested in. It is designed so that you could do a query like
`select * from comment` and it would properly suppress any and all
comments that should not be publicly viewable. To accomplish this, the
view's WHERE clause includes 9 correlated subqueries against 8 other
tables.

Most uses of the comment table would only be interested in the comments
that go with rows from a single "entity" table (for example `ipblocks`).
Having additional views that are restricted to a single entity type
should provide some performance improvements to queries that know to use
a specialized view.

New views added:
* comment_image
* comment_filearchive
* comment_ipblocks
* comment_oldimage
* comment_protected_titles
* comment_recentchanges
* comment_revision
* comment_logging

Each of these new views is a subset of the WHERE criteria for the
existing logging view limited to a single entity type. This means that
`SELECT * FROM comment_ipblocks` will only return rows from
the underlying comment table which are referenced by
ipblocks.ipb_reason_id where ipblocks.ipb_deleted is also 0.

Bug: T224850
Change-Id: I951c5d3c32a9e645059033ea333934e8de1c92cf