lvm_support: Increase timeout
When a disk failes the lvm commands take a lot longer to execute.
however they do eventually execute and values can be populated correctly.
This CR increase the timeout to 60 seconds.  We also add the
`noheadings` switch to one of the vgs commands to drop the need for the
regex math.


Bug: T223938
Change-Id: I2450ad934254237146a676098aa859e6d16497a1