openstack: drop hiera key profile::openstack::base::monitoring_host
We can use the more generic 'prometheus_nodes' instead. Propagate this hiera
key down to other related profiles, which were using a placeholder which can
be now replaced with 'prometheus_nodes' as well.

Since the new hiera key is an Array, we need to do some magic to convert
to the ferm syntax.

Also, we need to properly set the hiera keys for the affected servers.
Only cloudcontrol servers in this case, since cloudservices nodes already
got those set up. Why including the hiera config in the hosts/ tree? because
the datacenter tree takes precedence over the role/ tree, so we can't use
the role/ tree.

Change-Id: Iff5c43f86afec87da2490decc8e0cc5c7d4626f4
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>