profile::kerberos::kadminserver: allow puppetmaster to rsync keytabs
The idea is to manually generate, for the moment, Kerberos keytabs on
the kadmind master/active, and then allow puppetmaster1001 to rsync
them in order to be able to add them to the puppet private repository.

In this way the puppet master will not need to be a kerberos client.

As extra layer of protection, the rsync module will require user/pass
to pull data from puppetmaster1001.

Bug: T212257
Change-Id: Icb9b255df18014872852e42f7e610f8d801baeee