icinga: Add a script to parse and query the status.dat file
This CR adds a new `icinga-status` tool to icinga which allows user to
query the icinga status.dat file to determine if a specific host has
outstanding alerts.  The intention is to use this in spicrack
to determine if a host is healthy enough to be re-pooled.

I will send a comment to this CR with output examples to avoid commit
length warnings.  One thing to consider is do we need all the service
state.  it is fairly trivial to trim this list of values down if we can
decided on what we want.

Change-Id: I03729c09b0b2cf0b3c07affcaae9324527451c3b