wiki replicas: Add specialized views of the "actor" table
The actor table is not aware of the type of entity that a given query is
interested in. It is designed so that you could do a query like `select
* from actor` and it would properly suppress any and all actors that
should not be publicly viewable. To accomplish this, the view's WHERE
clause includes 9 correlated subqueries against other tables.

Most uses of the actor table would only be interested in the actors that
go with rows from a single "entity" table (for example `revision`).
Having additional views that are restricted to a single entity type
should provide some performance improvements to queries that know to use
a specialized view.

New views added:
* actor_archive
* actor_filearchive
* actor_image
* actor_ipblocks
* actor_logging
* actor_oldimage
* actor_recentchanges
* actor_revision
* actor_user

Each of these new views is a subset of the WHERE criteria for the
existing actor view limited to a single entity type. As an example, this
means that `SELECT * FROM actor_ipblocks` will only return rows from the
underlying actor table which are referenced by ipblocks.ipb_by_actor
where ipblocks.ipb_deleted is also 0.

This patch also adds allowed_logtypes filtering to both the `actor` and
`actor_logging` views to prevent side channel leaking of metadata
related to hidden log events.

Bug: T224850
Change-Id: Ic853bb8a3ea191f41f3fb0c490eda5f0594888f6