conftool::scripts: add a safe-service-restart script
This script picks up where pooler-loop was headed to create a more
integrated solution to safely restart a service that's under a conftool
group.

This script will:
* depool a node from the loadbalancer, for any selected lb pools
* Restart the selected local services
* repool only the services that were previously pooled

The advantage over the use of pooler-loop is we can now leave previously
depooled or inactive nodes in their previous state, preventing a repool
of a machine with faulty hardware for example.

The script also takes care of ensuring the server is actually depooled
by the load balancers, and then that it's properly repooled afterwards
before returning. This means that, if used with cumin, it will prevent
progressing to new machines until the current ones are properly dealt
with.

Future work might include adding possible concurrency limits so that we
prevent a restart to happen on all servers at around the same
time (although it should be noted that pybal will DTRT and thus there is
a sort-of concurrency protection built into the mechanism described above).

Bug: T224857
Bug: T104352
Change-Id: Iacf0aecce58f251f7ec18624c5b0268535a82b3f