systemd::timer::job: always provision NRPE plugin
* Move provisioning of
  /usr/local/lib/nagios/plugins/check_systemd_unit_status to a new
  ::systemd::timer::nrpe_plugin class
* Require ::systemd::timer::nrpe_plugin when $monitoring_enabled is true
  regardless of $ensure state

The systemd::timer::job define's Prior behavior was to conditionally
include the NRPE plugin inline here with a `if !defined(File[...]){...}`
construct and to follow $ensure state of the containing define. This
causes a non-deterministic state when a host has multiple
::systemd::timer::job defines applied with mixed $ensure state as the
first one in the graph will win and either ensure=>present or
ensure=>absent the plugin script. Let's be safe and just provision the
NRPE plugin for any $ensure state.

Bug: T225268
Change-Id: I66c07f705ded3220e307a2552048f54eb6a8d7fa