dologmsg: move this little script out of toolforge profile
This is so it is easier to make this something any VPS project can use
in the future.  It currently is linked to the Toolforge tools.* usernames.
In order to adapt it to generic VPS use, it will need that changed.

At that point, it can perhaps be included in an independent role to
select in Horizon for VPS users.

Change-Id: I0d5d14d5f2a546b83804eaa2a3d7905264dc5fda