dumps distribution: enable the service for nfs to start on reboot
This is somewhat controversial, but we decided it isn't bad.  This won't
break the failover setup on the dumps cluster if it is running.
However, it was a spontaneous reboot that triggered the interest in it.
We would never have known about the reboot had this setting been active.

We are now hoping that it will page on such a reboot anyway.

Bug: T217474
Change-Id: I39eae16a891b1cda3046ccc8ce771aff6d2ad860