ipmi - pxe: Ensure ipmi is not overriding the boot order
To calculate the what the ipmi boot deveice is i have created a new fact
`ipmi_chassis`.  This is produced by parsing the following command
  `sudo ipmi-chassis-config --checkout`
currently I have only added some of the values fro the output to the
fact.  If tjere are more you would like to see or if someof the ones
included are of no use please let me know.  The fact currently looks
like the following:

  % sudo facter -p ipmi_chassis
  {
    front_panel => {
      standby_button_enabled => true,
      reset_button_enabled => true,
      power_button_enabled => true
    },
    power => {
      restore_policy => "Restore_State_AC_Apply"
    },
    boot_flags => {
      persistent => false,
      device => "HARD-DRIVE",
      console_redirection => "BIOS-SETTING"
    }
  }

Change-Id: I539cc9a36a1ea50d8680aef4c2aa891300caf7ff