acme_chief: Introduce the concept of shared certificates
This commit implements the concept of shared acme-chief certificates.

The soon to be added ncredir puppet module needs to fetch details
of the non canonical certificates to configure TLS termination properly.
This is achieved by defining those certificates as shared certificates
on common.yaml, allowing acme-chief and other puppet modules to fetch
the details of those certificates.

Bug: T133548
Change-Id: Ie1efb19d37513d4fd5a07acee7b23c0bab1a0ea5