toolforge: k8s: etcd: don't wrap profile::etcd, and use base etcd v3 directly
Use the etcd v3 base class directly instead of wrapping profile::etcd.
In Toolforge, we don't need many of the things from profile::etcd,
and also this should help simplify the relationship between WMCS/Prod in this
puppet code.

Before this patch we were using an approach like this:
 * toolforge k8s etcd profile --> prod etcd profile --> etcd base class

After this patch, this is the appraoch instead:
 * toolforge k8s etcd profile --> etcd v3 base class

We can drop several hiera keys that we no longer require in this profile.

Bug: T226098
Change-Id: I8d07adb18166fdcf5e80d7493e3326184cfd3829
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>