peopleweb: Remove php module from httpd
This doesn't work because by default libapache2-mod-php7.0 ships
a config file that explicitly disables it for the only content
we serve on this host: user directories.

    <IfModule mod_userdir.c>
        <Directory /home/*/public_html>
            php_admin_flag engine Off
        </Directory>

Given it has been broken for last 7 months (possibly longer, maybe
on php 5 as well), just remove it.

Serving static files is convenient, but serving web apps and
dynamic scripts here is probably something semi-complicated to
support long-term regardless. Much better to use Toolforge for
that, where users can pick their PHP version, and is better
guarded and monitored for potential security issues.

Change-Id: I2c4b957014fde8840cb7c6c4f9993e29e222265a