toolforge: k8s: etcd: refresh server with certs changes
In WMCS switching puppetmasters is an operation that is more common than in
the prod realm. The etcd service is using puppet certs, and we had issues in
the past when we switched puppetmasters and there was a big mess of nothing
working in toolforge k8s/etcd.

One approach to prevent this mess is restart the etcd daemon if the
certificate changes, but puppet agent certs are not in the catalog, so let's do
this trick:

 * copy the puppet cert to the etcd ssl dir, therefore creating a File resource
 in the catalog
 * let's use this copied cert for the etcd config, now that we can notify on
 changes of this new File resource

Using this trick, we no longer need the etcd user to belong to the puppet
group.

Bug: T169287
Bug: T169287
Change-Id: Ie34ea262bce95ab483957385fda13e5bc17a1506
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>