profile::lvs::realserver: introduce ability to use safe-service-restart
While fixing some errors in safe_service_restart itself, and some
logical errors in profile::lvs::realserver.

Also use it on the canary appservers for now. We're here using a
different logic than before - any lvs pool has one or more corresponding
services that are needed to serve that pool. So when restarting a
service, we just depool any pool that service is necessary for.

I will update the class documentation and de-uglify the code once the
transition is completed everywhere.

Change-Id: I8237f0484aef251dfd224f2e93201051ddebfa7d