toolforge: correct a bunch of the apilb profile
The initial scribble I put up was incorrect in the template handling.
This should work correctly with the haproxy module as long as the patch
that makes monitoring more optional is ok.

If not, it'll be just as well to simply install the package and
collapse the config a bit.

Bug: T215531
Change-Id: I3202a9d6fd27f36a2eaaf06b1d2f36790130310f