cache: add cp3043 back to the text cluster
We moved cp3043 from text to upload in 6892cb2 as a way to alleviate
recurring Varnish scalability issues. Now that upload@esams is fully
converted to ATS, give the host back to the text cluster. This brings
the total count of text@esams hosts from 6 to 7, while decreasing upload
from 11 (it should be 12 but cp3037 is dead) to 10.

See if this helps with the recent fetch failures increase in text@esams.

Bug: T226375
Change-Id: I1e4546e7568dda2ee13da5a0650849b7ba36d582